+++ front/app/components/admin/Graphs/BarChart/BarChart.test.tsx
@@ -4,9 +4,6 @@ import BarChart from './';
 import { NilOrError } from 'utils/helperUtils';
 import { legacyColors } from '../styling';
 
-jest.mock('services/appConfiguration');
-jest.mock('utils/cl-intl');
-
 type Row = { name: string; value: number };
 
 const getEmptyData = (): Row[] | NilOrError => null;