+++ front/app/components/admin/ProjectableHeaderBgUploader/index.tsx
@@ -5,7 +5,11 @@ import { UploadFile } from 'typings';
 import ImagesDropzone from 'components/UI/ImagesDropzone';
 
 import { convertUrlToUploadFile } from 'utils/fileUtils';
-import { HEADER_BG_ASPECT_RATIO } from 'services/projects';
+import {
+  PROJECTABLE_HEADER_BG_ASPECT_RATIO,
+  PROJECTABLE_HEADER_BG_ASPECT_RATIO_HEIGHT,
+  PROJECTABLE_HEADER_BG_ASPECT_RATIO_WIDTH,
+} from 'services/projects';
 import ImageCropperContainer from 'components/admin/ImageCropper/Container';
 import ImageInfoTooltip from 'components/admin/ImageCropper/ImageInfoTooltip';
 import { SectionField, SubSectionTitle } from 'components/admin/Section';
@@ -17,7 +21,7 @@ interface Props {
   onImageChange: (newImageBase64: string | null) => void;
 }
 
-const HeaderBgInput = ({ imageUrl, onImageChange }: Props) => {
+const ProjectableHeaderBgUploader = ({ imageUrl, onImageChange }: Props) => {
   const [headerBg, setHeaderBg] = useState<UploadFile | null>(null);
   useEffect(() => {
     (async () => {
@@ -54,7 +58,8 @@ const HeaderBgInput = ({ imageUrl, onImageChange }: Props) => {
           <ImageCropperContainer
             image={headerBg}
             onComplete={onImageChange}
-            aspect={HEADER_BG_ASPECT_RATIO}
+            aspectRatioWidth={PROJECTABLE_HEADER_BG_ASPECT_RATIO_WIDTH}
+            aspectRatioHeight={PROJECTABLE_HEADER_BG_ASPECT_RATIO_HEIGHT}
             onRemove={handleImageRemove}
           />
         </Box>
@@ -64,7 +69,7 @@ const HeaderBgInput = ({ imageUrl, onImageChange }: Props) => {
           acceptedFileTypes={{
             'image/*': ['.jpg', '.jpeg', '.png', '.gif'],
           }}
-          imagePreviewRatio={1 / HEADER_BG_ASPECT_RATIO}
+          imagePreviewRatio={1 / PROJECTABLE_HEADER_BG_ASPECT_RATIO}
           onAdd={handleImageAdd}
           onRemove={handleImageRemove}
         />
@@ -73,4 +78,4 @@ const HeaderBgInput = ({ imageUrl, onImageChange }: Props) => {
   );
 };
 
-export default HeaderBgInput;
+export default ProjectableHeaderBgUploader;