+++ front/app/modules/commercial/smart_groups/components/UserFilterConditions/ValueSelector/NumberValueSelector.tsx
@@ -4,7 +4,7 @@ import { Input } from '@citizenlab/cl2-component-library';
 
 type Props = {
   value: string;
-  onChange: (string) => void;
+  onChange: (numberValue: number) => void;
 };
 
 interface State {}